telegram bot: rebuild modules tree
