+++ telegram_bot/src/commands/__init__.py
@@ -1,0 +1,3 @@
+from . import start
+from . import stop
+from . import help
