+++ telegram_bot/src/constants.py
@@ -1,11 +1,3 @@
 # components
 CONFIRM_BUTTON_TEXT: str = 'OK'
 CATALOGUE_CHECKBOX_EMOJI: str = ':ballot_box_with_check:'
-
-# commands
-HELP_MESSAGE: str = '\
-/start to create new search. \n\
-/stop to stop searching and delete created search. \n\
-\n\
-All bot\'s info You can find in the bot description.\
-'
